Avoid copying identical files

+++ bootstrap.sh
@@ -7,16 +7,25 @@ if [[ $yn != 'Y' ]]; then
 fi
 
 for f in $(find . -type f ! -path "./.git/*" ! -name "bootstrap.sh" ! -name "install.sh" ! -name "*.md" ! -name ".gitignore" ! -name ".editorconfig"); do
-    echo "Copying ${f:2}"
 
+    # Create parent directory
     d="${$(dirname $f):2}"
     if [[ d != "" ]]; then
         mkdir -p "$HOME/$d"
     fi
+
     fname="$HOME/${f:2}"
-    cp $f "$fname"
-    if [[ "$fname" == *.sh ]]; then
-        chmod +x "$fname"
+
+    # Copy
+    if cmp -s "$f" "$fname"; then
+        echo "Skipped ${f:2}"
+    else
+        echo "Copying ${f:2}"
+        cp $f "$fname"
+
+        if [[ "$fname" == *.sh ]]; then
+            chmod +x "$fname"
+        fi
     fi
 
 done
